refactor: reduce RPC requests on polls
