+++ src/services/contracts/DelegateRegistry.ts
@@ -1,9 +1,9 @@
-import type { JsonRpcSigner } from '@ethersproject/providers'
+import type { JsonRpcProvider } from '@ethersproject/providers'
 
 import { DELEGATE_REGISTRY_ADDRESS } from '@/config/constants'
 import { DelegateRegistry__factory } from '@/types/contracts/delegate-registry'
 import type { DelegateRegistry } from '@/types/contracts/delegate-registry/DelegateRegistry'
 
-export const getDelegateRegistryContract = (signer: JsonRpcSigner): DelegateRegistry => {
-  return DelegateRegistry__factory.connect(DELEGATE_REGISTRY_ADDRESS, signer)
+export const getDelegateRegistryContract = (provider: JsonRpcProvider): DelegateRegistry => {
+  return DelegateRegistry__factory.connect(DELEGATE_REGISTRY_ADDRESS, provider.getSigner())
 }
