+++ src/utils/__tests__/claim.test.ts
@@ -2,7 +2,7 @@ import { hexZeroPad, parseEther } from 'ethers/lib/utils'
 
 import { getAirdropInterface } from '@/services/contracts/Airdrop'
 import { createClaimTxs } from '@/utils/claim'
-import type { Vesting } from '@/hooks/useSafeTokenAllocation'
+import type { Vesting } from '@/hooks/useVestingData'
 
 const airdropInterface = getAirdropInterface()
