+++ src/utils/claim.ts
@@ -4,7 +4,7 @@ import type { BaseTransaction } from '@gnosis.pm/safe-apps-sdk/dist/src/types/sd
 import { getVestingTypes } from '@/utils/vesting'
 import { getAirdropInterface } from '@/services/contracts/Airdrop'
 import { splitAirdropAmounts } from '@/utils/airdrop'
-import type { Vesting } from '@/hooks/useSafeTokenAllocation'
+import type { Vesting } from '@/hooks/useVestingData'
 
 const airdropInterface = getAirdropInterface()
